+++ src/collectors/mod.rs
@@ -3,9 +3,13 @@ mod battery;
 mod disk;
 mod network;
 mod temperature;
+mod process;
+mod hosts;
 
 pub use system::SystemInfo;
 pub use battery::BatteryInfo;
 pub use disk::DisksInfo;
 pub use network::NetworkInfo;
 pub use temperature::TemperatureInfo;
+pub use process::ProcessInfo;
+pub use hosts::HostsInfo;